Add "main" entrypoint to each package.json file
The "module" entrypoint appears to be sufficient for webpack/rollup, but
npm itself doesn't understand it, so, for example, if you want to run
tests with jest, it won't recognize any of these as modules that can be
imported from.

I've tested this by making this change under node_modules/ in a package
of mine, and this appeared to fix the problem.